fix(WEM-2048): fix issue with "Add to cart" button on category page
The "Add to cart" button on the category page was not working correctly when adding a product to the cart from the "Recently Ordered" section. This was due to incorrect information being saved in the draft, which prevented the draft from being opened in the editor.

This commit fixes the issue by updating the information in the draft and resolving the "draft cannot be opened by you" error message.

Fixes #WEM-2048